+++ docs/CHANGELOG.md
@@ -86,6 +86,8 @@ Version 1.2.0 (Willow): Not yet released
 - Add `When_Captured` event
 - Add `Owner_Gate` event
 - Allow using `Edit Links` window in level editor to link objects to MP respawn points
+- Support rule presets and aliases in ADS dedicated servers
+- Allow preset support to `vote level` and `vote match` chat commands
 
 [@is-this-c](https://github.com/is-this-c)
 - Simplify code that fixed broken events if their delay parameter was set